Update Blogpost_text.md
Hi Julien! Good work in reflecting on the workshop!
As mentioned - we can of course publish this with your credentials on the A2P blog. Additionally - if you can also get the SFB people share it in their staff newsletter would be great. Let me know when and if I should put it up - please double check if we can name the SFB and Bochum University publicly as clients in the article - otherwise we can also publish it anonymised.

Happy Holidays!!